requirements.txt
pypiwin32; sys_platform == 'win32'
-e git+ssh://git@gitlab.crosscloud.me/CrossCloud/bushn.git@v1.0.10#egg=bushn
atomicwrites==1.1.5
better-exceptions==0.1.6
blinker==1.4.0
bumpversion==0.5.3
filelock==2.0.6
macfsevents==0.7; sys_platform == 'darwin'
mock==2.0.0
pylint==1.5.6
pytest-mock==1.6.0
pytest-cov==2.4.0
pytest==3.0.7
python-dateutil==2.5.3
pytz==2017.2
raven==6.0.0
requests_oauthlib==0.7.0
requests-mock==1.3.0
requests-ntlm==1.0.0
requests==2.17.3
termcolor==1.1.0
watchdog==0.8.3
